Added the feature for the teacher to add Tasks

+++ php/src/app/admin/deletefile.php
@@ -6,9 +6,14 @@ if(!isset($_SESSION["username"])) {
     exit();
 }
 
+if($_SESSION["permission"] != "admin") {
+    header("Location: ../../index.php");
+    exit();
+}
+
 if(isset($_POST["submit"])) {
     $file = $_POST["file"];
-    $fileDestination = '../users/' . $_SESSION["username"] . '/' . $file;
+    $fileDestination = '../users/#Task' . '/' . $file;
     unlink($fileDestination);
     header("Location: dashboard.php?delete=success");
 }
